requirements.txt
python-telegram-bot==20.7
pandas==2.1.0
ta==0.10.2
websocket-client==1.6.1
requests==2.31.0
deriv-api==1.0.4